+++ app/views/pdfs/show.html.erb
@@ -18,7 +18,7 @@
 <%= render('pdfs/risk_summary', risks: escort.risks) %>
 <%= render('pdfs/healthcare_needs', healthcare: escort.healthcare) %>
 <%= render('pdfs/offences', offences: escort.offences) %>
-<%= render('pdfs/handover_details', handover: escort.handover) %>
+<%= render('pdfs/handover_details') %>
 <%= render('pdfs/record_of_handover') %>
 <%= render('pdfs/checks_and_events_initial') %>
 <%= render('pdfs/checks_and_events_continuation') %>
